client: Refactor syntax highlighting for read-only text into service
